Add api functions and types to handle offer management
