+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from "./user.types";
 export * from "./caretaker.types";
 export * from "./pagination.types";
+export * from "./offer.types";
