+++ src/types/offer.types.ts
@@ -1,0 +1,47 @@
+export type OfferConfigurationDTO = {
+  description: string;
+  dailyPrice: number;
+  selectedOptions: Record<string, string[]>;
+}
+
+export type OfferConfigurationWithId = OfferConfigurationDTO & {
+  id: number;
+}
+
+export type AvailabilityDTO = {
+  availableFrom: string;
+  availableTo: string;
+}
+
+export type SetAvailabilityDTO = {
+  offerIds: number[];
+  availabilityRanges: AvailabilityDTO[];
+}
+
+export type OfferDTO = {
+  description: string;
+  animal: {
+    animalType: string;
+  };
+  offerConfigurations: OfferConfigurationDTO[];
+  animalAmenities: string[];
+}
+
+export type OfferDTOWithId = Omit<OfferDTO, "offerConfigurations"> & {
+  id: number;
+  offerConfigurations: OfferConfigurationWithId[];
+  availabilities: AvailabilityDTO[];
+}
+
+export type AnimalSex = "MALE" | "SHE";
+export type AnimalSize = "SMALL" | "MEDIUM" | "BIG"; 
+
+export type OfferConfiguration = {
+  attributes?: {
+    SIZE?: AnimalSize[];
+    SEX?: AnimalSex[];
+  };
+  minPrice?: number;
+  maxPrice?: number;
+  amenities?: string[];
+};
